feat: extract most of checksum to shared base class
