+++ src/crc32c/python.py
@@ -15,6 +15,8 @@
 import array
 import struct
 
+from crc32c._checksum import CommonChecksum
+
 
 def extend(crc, chunk):
     """Update an existing CRC checksum with new chunk of data.
@@ -49,7 +51,7 @@ def value(chunk):
     return c._crc
 
 
-class Checksum(object):
+class Checksum(CommonChecksum):
     """Hashlib-alike helper for CRC32C operations.
 
     Args:
@@ -78,52 +80,6 @@ class Checksum(object):
             table_poly = _TABLE[(b ^ self._crc) & 0xFF]
             self._crc = table_poly ^ ((self._crc >> 8) & 0xFFFFFFFF)
         self._crc = self._crc ^ 0xFFFFFFFF
-
-    def digest(self):
-        """Big-endian order, per RFC 4960.
-
-        See: https://cloud.google.com/storage/docs/json_api/v1/objects#crc32c
-
-        Returns:
-            bytes: An eight-byte digest string.
-        """
-        return struct.pack(">L", self._crc)
-
-    def hexdigest(self):
-        """Like :meth:`digest` except returns as a bytestring of double length.
-
-        Returns
-            bytes: A sixteen byte digest string, contaiing only hex digits.
-        """
-        return "{:08x}".format(self._crc).encode("ascii")
-
-    def copy(self):
-        """Create another checksum with the same CRC32C value.
-
-        Returns:
-            Checksum: the new instance.
-        """
-        clone = self.__class__()
-        clone._crc = self._crc
-        return clone
-
-    def consume(self, stream, chunksize):
-        """Consume chunks from a stream, extending our CRC32 checksum.
-
-        Args:
-            stream (BinaryIO): the stream to consume.
-            chunksize (int): the size of the read to perform
-
-        Returns:
-            Generator[bytes, None, None]: Tterable of the chunks read from the
-            stream.
-        """
-        while True:
-            chunk = stream.read(chunksize)
-            if not chunk:
-                break
-            self._crc = self.update(chunk)
-            yield chunk
 
 
 # fmt:off
